requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
websockets==12.0
redis==5.0.0
httpx==0.26.0
cryptography==41.0.7
pynacl==1.5.0
qrcode==7.4.2
pytest==8.0.0